[AIRFLOW-5498] Move GCP Discovery hook to core (#6121)
